==> 0 - Obtendo os clusters iniciais.ipynb <==
import pandas as pd
import numpy as np

from sklearn.preprocessing import Normalizer
from sklearn.cluster import KMeans

base = pd.read_csv(r'Dados/general.csv')

# Limpando colunas desnecessárias
base.drop(['CUST_ID', 'TENURE'], axis=1, inplace=True)


# Verificando dados faltantes
base.isnull().sum()

# Completando os valores faltantes com a mediana
base.fillna(base.median(), inplace=True)

# Normalizando os valores
valores_normalizados = Normalizer().fit_transform(base.values)

# Obtendo os clusters
kmeans = KMeans(n_clusters=5, n_init=10, max_iter=300)
prediction = kmeans.fit_predict(valores_normalizados)

# Exportando dados normalizados e previsões iniciais
np.savetxt('Dados/valores_normalizados.csv', valores_normalizados, delimiter=',')
np.savetxt('Dados/clusters_iniciais.csv', prediction, delimiter=',')
